kernel: fix deadlock when stopping streaming acquisition
While a block acquisition is running the ZIO flag HW_BUSY cannot be
cleared because I disable the interrupts before completing the acquisition
and this flag is cleared only when the block acquisition is over.

This patch introduce the stopping flag which signal to the irq handler
that they must stop the acquisition as soon as they receive the acquisition
complete interrupt, or before starting a new acquisition.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>